+++ src/LowerTriangularMatrices/lower_triangular_array.jl
@@ -211,7 +211,6 @@ end
 
 # l,m sph "matrix-style"  indexing with integer + other indices
 @inline function Base.getindex(L::LowerTriangularArray{T,N}, I::Vararg{Any, M}) where {T, N, M}
-    @boundscheck M == N+1 || throw(BoundsError(L, I))
     i, j = I[1:2]
     @boundscheck (0 < i <= L.m && 0 < j <= L.n) || throw(BoundsError(L, (i, j)))
     # to get a zero element in the correct shape, we just take the zero element of some valid element,
@@ -631,10 +630,10 @@ function Base.similar(
     return LowerTriangularArray{T, N, ArrayType{T,N}}(undef, size(L; as=Matrix))
 end
 
-function GPUArrays.backend(
-    ::Type{LowerTriangularArray{T, N, ArrayType}}
+function KernelAbstractions.get_backend(
+    a::LowerTriangularArray{T, N, ArrayType}
 ) where {T, N, ArrayType <: GPUArrays.AbstractGPUArray}
-    return GPUArrays.backend(ArrayType)
+    return KernelAbstractions.get_backend(a.data)
 end
 
 Adapt.adapt_structure(to, L::LowerTriangularArray) =