+++ src/models/primitive_dry.jl
@@ -34,8 +34,10 @@ $(TYPEDFIELDS)"""
     SUW,    # <:AbstractSurfaceWind,
     SH,     # <:AbstractSurfaceHeatFlux,
     CV,     # <:AbstractConvection,
+    OD,     # <:AbstractOpticalDepth,
     SW,     # <:AbstractShortwave,
     LW,     # <:AbstractLongwave,
+    SP,     # <:AbstractStochasticPhysics,
     TS,     # <:AbstractTimeStepper,
     ST,     # <:SpectralTransform{NF},
     IM,     # <:AbstractImplicit,
@@ -77,8 +79,10 @@ $(TYPEDFIELDS)"""
     surface_wind::SUW = SurfaceWind(spectral_grid)
     surface_heat_flux::SH = SurfaceHeatFlux(spectral_grid)
     convection::CV = DryBettsMiller(spectral_grid)
+    optical_depth::OD = ZeroOpticalDepth(spectral_grid)
     shortwave_radiation::SW = NoShortwave(spectral_grid)
     longwave_radiation::LW = JeevanjeeRadiation(spectral_grid)
+    stochastic_physics::SP = StochasticallyPerturbedParameterizationTendencies(spectral_grid)
     
     # NUMERICS
     time_stepping::TS = Leapfrog(spectral_grid)
@@ -126,11 +130,14 @@ function initialize!(model::PrimitiveDry; time::DateTime = DEFAULT_DATE)
     initialize!(model.boundary_layer_drag, model)
     initialize!(model.temperature_relaxation, model)
     initialize!(model.vertical_diffusion, model)
+    initialize!(model.convection, model)
+    initialize!(model.optical_depth, model)
     initialize!(model.shortwave_radiation, model)
     initialize!(model.longwave_radiation, model)
     initialize!(model.surface_thermodynamics, model)
     initialize!(model.surface_wind, model)
     initialize!(model.surface_heat_flux, model)
+    initialize!(model.stochastic_physics, model)
 
     # initial conditions
     prognostic_variables = PrognosticVariables(spectral_grid, model)
@@ -139,15 +146,16 @@ function initialize!(model::PrimitiveDry; time::DateTime = DEFAULT_DATE)
     clock.time = time       # set the current time
     clock.start = time      # and store the start time
     
-    diagnostic_variables = DiagnosticVariables(spectral_grid)
+    diagnostic_variables = DiagnosticVariables(spectral_grid, model)
     
     # particle advection
     initialize!(model.particle_advection, model)
     initialize!(prognostic_variables.particles, model)
 
     # initialize ocean and land
-    initialize!(prognostic_variables.ocean, time, model)
+    initialize!(prognostic_variables.ocean, prognostic_variables, diagnostic_variables, model)
     initialize!(prognostic_variables.land,  prognostic_variables, diagnostic_variables, model)
 
+    # pack prognostic, diagnostic variables and model into a simulation
     return Simulation(prognostic_variables, diagnostic_variables, model)
 end