+++ src/models/primitive_wet.jl
@@ -39,8 +39,10 @@ $(TYPEDFIELDS)"""
     EV,     # <:AbstractSurfaceEvaporation,
     LSC,    # <:AbstractCondensation,
     CV,     # <:AbstractConvection,
+    OD,     # <:AbstractOpticalDepth,
     SW,     # <:AbstractShortwave,
     LW,     # <:AbstractLongwave,
+    SP,     # <:AbstractStochasticPhysics,
     TS,     # <:AbstractTimeStepper,
     ST,     # <:SpectralTransform{NF},
     IM,     # <:AbstractImplicit,
@@ -88,8 +90,10 @@ $(TYPEDFIELDS)"""
     surface_evaporation::EV = SurfaceEvaporation(spectral_grid)
     large_scale_condensation::LSC = ImplicitCondensation(spectral_grid)
     convection::CV = SimplifiedBettsMiller(spectral_grid)
+    optical_depth::OD = ZeroOpticalDepth(spectral_grid)
     shortwave_radiation::SW = NoShortwave(spectral_grid)
     longwave_radiation::LW = JeevanjeeRadiation(spectral_grid)
+    stochastic_physics::SP = StochasticallyPerturbedParameterizationTendencies(spectral_grid)
     
     # NUMERICS
     time_stepping::TS = Leapfrog(spectral_grid)
@@ -142,29 +146,32 @@ function initialize!(model::PrimitiveWet; time::DateTime = DEFAULT_DATE)
     initialize!(model.vertical_diffusion, model)
     initialize!(model.large_scale_condensation, model)
     initialize!(model.convection, model)
+    initialize!(model.optical_depth, model)
     initialize!(model.shortwave_radiation, model)
     initialize!(model.longwave_radiation, model)
     initialize!(model.surface_thermodynamics, model)
     initialize!(model.surface_wind, model)
     initialize!(model.surface_heat_flux, model)
     initialize!(model.surface_evaporation, model)
+    initialize!(model.stochastic_physics, model)
 
     # initial conditions
     prognostic_variables = PrognosticVariables(spectral_grid, model)
-    initialize!(prognostic_variables,model.initial_conditions, model)
-    (;clock) = prognostic_variables
+    initialize!(prognostic_variables, model.initial_conditions, model)
+    (; clock) = prognostic_variables
     clock.time = time       # set the current time
     clock.start = time      # and store the start time
 
-    diagnostic_variables = DiagnosticVariables(spectral_grid)
+    diagnostic_variables = DiagnosticVariables(spectral_grid, model)
     
     # particle advection
     initialize!(model.particle_advection, model)
     initialize!(prognostic_variables.particles, model)
 
     # initialize ocean and land
-    initialize!(prognostic_variables.ocean, time, model)
+    initialize!(prognostic_variables.ocean, prognostic_variables, diagnostic_variables, model)
     initialize!(prognostic_variables.land,  prognostic_variables, diagnostic_variables, model)
 
+    # pack prognostic, diagnostic variables and model into a simulation
     return Simulation(prognostic_variables, diagnostic_variables, model)
 end