+++ src/physics/shortwave_radiation.jl
@@ -1,6 +1,11 @@
 abstract type AbstractRadiation <: AbstractParameterization end
 abstract type AbstractShortwave <: AbstractRadiation end
 
+function get_nbands(R::AbstractRadiation)
+    hasfield(typeof(R), :nbands) && return R.nbands
+    return 0
+end
+
 # function barrier for all AbstractShortwave
 function shortwave_radiation!(column::ColumnVariables, model::PrimitiveEquation)
     shortwave_radiation!(column, model.shortwave_radiation, model)
@@ -35,7 +40,14 @@ function shortwave_radiation!(
     (; cos_zenith, albedo) = column
     (; solar_constant) = planet
 
-    # reduce strength by 1/4 as this currently just hits the air temperature in lowermost
-    # layer directly and not through a skin temperature
-    column.flux_temp_upward[end] += ((1-albedo) * solar_constant * cos_zenith)/4
+    # transparent = optical thickness of zero, no vertical changes in flux
+    # this will sum up to zero in every layer (=transparent) but yields
+    # a non-zero net flux at the surface 
+    column.flux_temp_downward .+= solar_constant * cos_zenith
+    column.flux_temp_upward .+= albedo * solar_constant * cos_zenith
+
+    # diagnostic: outgoing shortwave radiation
+    column.outgoing_shortwave_radiation = albedo * solar_constant * cos_zenith
+
+    return nothing
 end