+++ test/runtests.jl
@@ -1,6 +1,12 @@
 using SpeedyWeather
 using Test
 
+FLAG_EXTENDED_TESTS = "extended_tests" in ARGS ? true : false
+
+if FLAG_EXTENDED_TESTS 
+    @info "Running extended test suite"
+end 
+
 # GENERAL
 include("utility_functions.jl")
 include("dates.jl")
@@ -16,6 +22,8 @@ include("kernelabstractions.jl")
 # SPECTRAL TRANSFORM
 include("spectral_transform.jl")
 include("spectral_gradients.jl")
+include("spectrum.jl")
+include("spectral_transform_ad_rules.jl") 
 
 # DYNAMICS
 include("diffusion.jl")
@@ -34,11 +42,13 @@ include("orography.jl")
 include("land_sea_mask.jl")
 include("ocean.jl")
 # include("thermodynamics.jl")
-# include("large_scale_condensation.jl")
+include("large_scale_condensation.jl")
 # include("convection.jl")
-# include("longwave_radiation.jl")
+include("optical_depth.jl")
+include("longwave_radiation.jl")
 # include("shortwave_radiation.jl")
 include("random_process.jl")
+include("stochastic_physics.jl")
 
 # INITIALIZATION AND INTEGRATION
 include("run_speedy.jl")